chore: cleanup decoders
