+++ src/network/os1/client/codec/game/OpLocDecoder.ts
@@ -6,14 +6,54 @@ import MessageDecoder from '#/network/client/codec/MessageDecoder.ts';
 import OpLoc from "#/network/client/model/game/OpLoc.js";
 
 export default class OpLocDecoder extends MessageDecoder {
-    opcode = 73;
     size = 6;
     limit = GameClientLimit.CLIENT;
+    op: number;
+    opcode: number;
+
+    constructor(op: number, opcode: number) {
+        super();
+        this.op = op;
+        this.opcode = opcode
+    }
 
     read(buf: Packet): ClientMessage {
-        const locId = buf.g2_alt2();
-        const x = buf.g2();
-        const z = buf.g2();
-        return new OpLoc(1, locId, x, z);
+        switch (this.op) {
+            case 1: {
+                const locId = buf.g2_alt2();
+                const x = buf.g2();
+                const z = buf.g2();
+                return new OpLoc(1, locId, x, z);
+            }
+            case 2: {
+                const z = buf.g2_alt3();
+                const x = buf.g2_alt3();
+                const locId = buf.g2_alt2();
+                return new OpLoc(2, locId, x, z);
+            }
+            case 3: {
+                const x = buf.g2_alt2();
+                const z = buf.g2_alt2();
+                const locId = buf.g2_alt3();
+                return new OpLoc(3, locId, x, z);
+            }
+            case 4: {
+                const x = buf.g2_alt2();
+                const z = buf.g2_alt3();
+                const locId = buf.g2_alt3();
+                return new OpLoc(4, locId, x, z);
+            }
+            case 5: {
+                const x = buf.g2();
+                const locId = buf.g2_alt1();
+                const z = buf.g2_alt2();
+                return new OpLoc(5, locId, x, z);
+            }
+            default:
+            {
+                console.error(`Unhandled 'OpLoc' op value: ${this.op}`);
+                process.exit(1);
+            }
+        }
     }
 }
