--- src/network/os1/client/codec/game/OpNpc4Decoder.ts
@@ -1,17 +1,0 @@
-import type Packet from '#/io/Packet.ts';
-
-import type ClientMessage from '#/network/client/ClientMessage.ts';
-import GameClientLimit from '#/network/client/codec/game/GameClientLimit.ts';
-import MessageDecoder from '#/network/client/codec/MessageDecoder.ts';
-import OpNpc from "#/network/client/model/game/OpNpc.js";
-
-export default class OpNpc4Decoder extends MessageDecoder {
-    opcode = 95;
-    size = 2;
-    limit = GameClientLimit.CLIENT;
-
-    read(buf: Packet): ClientMessage {
-        const npcId = buf.g2_alt1();
-        return new OpNpc(4, npcId);
-    }
-}
