--- src/network/os1/client/codec/game/OpPlayer3Decoder.ts
@@ -1,17 +1,0 @@
-import type Packet from '#/io/Packet.ts';
-
-import type ClientMessage from '#/network/client/ClientMessage.ts';
-import GameClientLimit from '#/network/client/codec/game/GameClientLimit.ts';
-import MessageDecoder from '#/network/client/codec/MessageDecoder.ts';
-import OpPlayer from "#/network/client/model/game/OpPlayer.js";
-
-export default class OpPlayer3Decoder extends MessageDecoder {
-    opcode = 102;
-    size = 2;
-    limit = GameClientLimit.CLIENT;
-
-    read(buf: Packet): ClientMessage {
-        const playerId = buf.g2_alt1();
-        return new OpPlayer(3, playerId);
-    }
-}
