--- src/network/os1/client/codec/game/OpPlayer6Decoder.ts
@@ -1,17 +1,0 @@
-import type Packet from '#/io/Packet.ts';
-
-import type ClientMessage from '#/network/client/ClientMessage.ts';
-import GameClientLimit from '#/network/client/codec/game/GameClientLimit.ts';
-import MessageDecoder from '#/network/client/codec/MessageDecoder.ts';
-import OpPlayer from "#/network/client/model/game/OpPlayer.js";
-
-export default class OpPlayer6Decoder extends MessageDecoder {
-    opcode = 111;
-    size = 2;
-    limit = GameClientLimit.CLIENT;
-
-    read(buf: Packet): ClientMessage {
-        const playerId = buf.g2_alt3();
-        return new OpPlayer(6, playerId);
-    }
-}
