+++ src/network/os1/client/prot/game/GameClientRepository.ts
@@ -10,7 +10,7 @@ import ClanDecoder from '../../codec/game/ClanDecoder.ts';
 import IfButtonDecoder from '../../codec/game/IfButtonDecoder.ts';
 import ClanHandler from '../../handler/game/ClanHandler.ts';
 import IfButtonHandler from '../../handler/game/IfButtonHandler.ts';
-import OpPlayer1Decoder from "#/network/os1/client/codec/game/OpPlayer1Decoder.js";
+import OpPlayerDecoder from "#/network/os1/client/codec/game/OpPlayerDecoder.js";
 import OpPlayerHandler from "#/network/os1/client/handler/game/OpPlayerHandler.js";
 import OpPlayer2Decoder from "#/network/os1/client/codec/game/OpPlayer2Decoder.js";
 import OpPlayer3Decoder from "#/network/os1/client/codec/game/OpPlayer3Decoder.js";
@@ -19,24 +19,12 @@ import OpPlayer5Decoder from "#/network/os1/client/codec/game/OpPlayer5Decoder.j
 import OpPlayer6Decoder from "#/network/os1/client/codec/game/OpPlayer6Decoder.js";
 import OpPlayer7Decoder from "#/network/os1/client/codec/game/OpPlayer7Decoder.js";
 import OpPlayer8Decoder from "#/network/os1/client/codec/game/OpPlayer8Decoder.js";
-import OpNpc1Decoder from "#/network/os1/client/codec/game/OpNpc1Decoder.js";
+import OpNpcDecoder from "#/network/os1/client/codec/game/OpNpcDecoder.js";
 import OpNpcHandler from "#/network/os1/client/handler/game/OpNpcHandler.js";
-import OpNpc2Decoder from "#/network/os1/client/codec/game/OpNpc2Decoder.js";
-import OpNpc3Decoder from "#/network/os1/client/codec/game/OpNpc3Decoder.js";
-import OpNpc4Decoder from "#/network/os1/client/codec/game/OpNpc4Decoder.js";
-import OpNpc5Decoder from "#/network/os1/client/codec/game/OpNpc5Decoder.js";
 import OpLocDecoder from "#/network/os1/client/codec/game/OpLocDecoder.js";
 import OpLocHandler from "#/network/os1/client/handler/game/OpLocHandler.js";
-import OpLoc2Decoder from "#/network/os1/client/codec/game/OpLoc2Decoder.js";
-import OpLoc3Decoder from "#/network/os1/client/codec/game/OpLoc3Decoder.js";
-import OpLoc4Decoder from "#/network/os1/client/codec/game/OpLoc4Decoder.js";
-import OpLoc5Decoder from "#/network/os1/client/codec/game/OpLoc5Decoder.js";
-import OpObj1Decoder from "#/network/os1/client/codec/game/OpObj1Decoder.js";
+import OpObjDecoder from "#/network/os1/client/codec/game/OpObjDecoder.js";
 import OpObjHandler from "#/network/os1/client/handler/game/OpObjHandler.js";
-import OpObj2Decoder from "#/network/os1/client/codec/game/OpObj2Decoder.js";
-import OpObj3Decoder from "#/network/os1/client/codec/game/OpObj3Decoder.js";
-import OpObj4Decoder from "#/network/os1/client/codec/game/OpObj4Decoder.js";
-import OpObj5Decoder from "#/network/os1/client/codec/game/OpObj5Decoder.js";
 import LocExamineDecoder from "#/network/os1/client/codec/game/LocExamineDecoder.js";
 import LocExamineHandler from "#/network/os1/client/handler/game/LocExamineHandler.js";
 import NpcExamineDecoder from "#/network/os1/client/codec/game/NpcExamineDecoder.js";
@@ -63,29 +51,29 @@ export default class GameClientRepository extends ClientRepository {
         this.bind(new ClientCheatDecoder(), new ClientCheatHandler())
         this.bind(new IfButtonDecoder(), new IfButtonHandler())
         this.bind(new ClanDecoder(), new ClanHandler())
-        this.bind(new OpPlayer1Decoder(), new OpPlayerHandler())
-        this.bind(new OpPlayer2Decoder(), new OpPlayerHandler())
-        this.bind(new OpPlayer3Decoder(), new OpPlayerHandler())
-        this.bind(new OpPlayer4Decoder(), new OpPlayerHandler())
-        this.bind(new OpPlayer5Decoder(), new OpPlayerHandler())
-        this.bind(new OpPlayer6Decoder(), new OpPlayerHandler())
-        this.bind(new OpPlayer7Decoder(), new OpPlayerHandler())
-        this.bind(new OpPlayer8Decoder(), new OpPlayerHandler())
-        this.bind(new OpNpc1Decoder(), new OpNpcHandler())
-        this.bind(new OpNpc2Decoder(), new OpNpcHandler())
-        this.bind(new OpNpc3Decoder(), new OpNpcHandler())
-        this.bind(new OpNpc4Decoder(), new OpNpcHandler())
-        this.bind(new OpNpc5Decoder(), new OpNpcHandler())
-        this.bind(new OpLocDecoder(), new OpLocHandler())
-        this.bind(new OpLoc2Decoder(), new OpLocHandler())
-        this.bind(new OpLoc3Decoder(), new OpLocHandler())
-        this.bind(new OpLoc4Decoder(), new OpLocHandler())
-        this.bind(new OpLoc5Decoder(), new OpLocHandler())
-        this.bind(new OpObj1Decoder(), new OpObjHandler())
-        this.bind(new OpObj2Decoder(), new OpObjHandler())
-        this.bind(new OpObj3Decoder(), new OpObjHandler())
-        this.bind(new OpObj4Decoder(), new OpObjHandler())
-        this.bind(new OpObj5Decoder(), new OpObjHandler())
+        this.bind(new OpPlayerDecoder(1, 246), new OpPlayerHandler())
+        this.bind(new OpPlayerDecoder(2, 146), new OpPlayerHandler())
+        this.bind(new OpPlayerDecoder(3, 102), new OpPlayerHandler())
+        this.bind(new OpPlayerDecoder(4, 78), new OpPlayerHandler())
+        this.bind(new OpPlayerDecoder(5, 117), new OpPlayerHandler())
+        this.bind(new OpPlayerDecoder(6, 111), new OpPlayerHandler())
+        this.bind(new OpPlayerDecoder(7, 119), new OpPlayerHandler())
+        this.bind(new OpPlayerDecoder(8, 145), new OpPlayerHandler())
+        this.bind(new OpNpcDecoder(1, 84), new OpNpcHandler())
+        this.bind(new OpNpcDecoder(2, 13), new OpNpcHandler())
+        this.bind(new OpNpcDecoder(3, 67), new OpNpcHandler())
+        this.bind(new OpNpcDecoder(4, 95), new OpNpcHandler())
+        this.bind(new OpNpcDecoder(5, 88), new OpNpcHandler())
+        this.bind(new OpLocDecoder(1, 73), new OpLocHandler())
+        this.bind(new OpLocDecoder(2, 90), new OpLocHandler())
+        this.bind(new OpLocDecoder(3, 133), new OpLocHandler())
+        this.bind(new OpLocDecoder(4, 83), new OpLocHandler())
+        this.bind(new OpLocDecoder(5, 56), new OpLocHandler())
+        this.bind(new OpObjDecoder(1, 135), new OpObjHandler())
+        this.bind(new OpObjDecoder(2, 179), new OpObjHandler())
+        this.bind(new OpObjDecoder(3, 76), new OpObjHandler())
+        this.bind(new OpObjDecoder(4, 220), new OpObjHandler())
+        this.bind(new OpObjDecoder(5, 19), new OpObjHandler())
         this.bind(new LocExamineDecoder(), new LocExamineHandler())
         this.bind(new NpcExamineDecoder(), new NpcExamineHandler())
         this.bind(new TileObjExamineDecoder(), new ObjExamineHandler())
